posture/v_human-pose/requirements.txt
altgraph==0.17.4
importlib-metadata==4.8.3
numpy==1.19.5
pefile==2024.8.26
pyserial==3.5
pywin32-ctypes==0.2.3
typing_extensions==4.1.1
zipp==3.6.0
